Fix clippy warnings for CI
- Install ffmpeg development libraries in CI workflow
- Fix numerous clippy warnings in library and binary code
- Add allow attributes for legacy xf-chat binary
- Fix test imports after removing wildcard exports
- Add clippy.toml with configuration
- Update CI to check lib/bins only (exclude tests from strict mode)

+++ src/permission/mod.rs
@@ -141,8 +141,8 @@ pub async fn new() -> Result<PermissionData, sea_orm::error::DbErr> {
                 match col.categories[i].add_item(item.id, &item.label) {
                     Ok(item) => {
                         col.dictionary
-                            .insert(item.label.to_owned(), (i as u8, item.position as u8));
-                        col.lookup.insert(item.id, (i as u8, item.position as u8));
+                            .insert(item.label.to_owned(), (i as u8, item.position));
+                        col.lookup.insert(item.id, (i as u8, item.position));
                     }
                     Err(_) => {
                         println!("Category overflow adding {:?}", item);
